Modify gofer/run.go to work with Titan.
This patch converts gofer/run.go to work with the Titan API. Attempts to share
code with iron-io/go were not entirely successful.

The following are used directly from iron-io/go. We'll need to make those open
source before we release titan:
- common/stats
- common/{environment,context,stats}.go

While I would have liked to stick to the Tasker and Logger interfaces defined
in the gofer runner, the dependence on *backend.Task and bson.ObjectId prevents
this.  In addition Code & CodeHistory are too tightly coupled to those
interfaces to directly use this in Titan. To be clear, it is not impossible to
share more code, but it has not been done in the interest of having a runner
quickly.  Sharing such code would require more functionality to move into the
swapi tasker, where it would maintain a mapping from backend.Task to
ContainerTask, and deal with histories and so on itself. Then it could be
folded into Titan. Unfortunately that would make this PR significantly larger
and harder to review.