Clean up job directory after job finished. (#117)
Imposes an additional contract on the RunResult to implement the io.Closer
interface so that the runner can call close on it after it is done
communicating with the remote API and clean up any resources that need to be
cleaned up. Added a default implementation for the Docker driver.