Refactor runner to use tasker and jobs as interfaces.
The current runner was a copy of the gorunner, with API changes that tightly
coupled it to Titan. Now that we know it works, this branch is a work in
progress to reconcile the two copies to just one. This first step makes it so
that the runner is not aware of Titan tasker and job as concrete types.

- Makes runner the top-level package that can be used by main.go in both titan and swapi.
- Moves titan runner's main.go to runner/cmd/runner.
- Moves titan tasker to runner/tasker.
- Moves the go:generate bits to tasker.
- Disables log upload support in titan until we can reconcile our io.ReadSeeker requirements with swagger generated io.File.