log16: one log api
previously, we were using log15 in certain places and logrus in certain
places. both were pretty deep in our shit, but differed a lot in style; i.e.
each one did things much differently than the other. this seems like a bad
idea for more reasons than uniformity of logs emitted, but also in the code
(aside, log levels were set on one correctly, but not the other...).

logrus was pretty deeply embedded in all of our shit, and our logs were worse
off for it. logrus doesn't have one standard api, instead it has 3, and makes
use of WithFields being mandatory rather than doing what one might expect in a
log line and using fields as fields. logrus doesn't seem very efficient but
somebody seems hellbent on using it, so in an attempt to compromise, this
makes a sane api for logging with only 5 options and then a thin wrapper
around logrus shoves things from the sane api into logrus' insane api in the
way that one might have expected in the first place. this will make the logs
cleaner, as we had a lot of logs where fields were being concatenated together
making them challenging to read and impossible for a computer to parse.

a side benefit of this is that swapping out to a different log package will be
really easy. after seeing logrus' guts, I'm inclined to switch to log15 or
glog (which are also plenty popular), but the logrus thing works for now if it
can get this through review -- if it comes up in profiles, we can reconsider
(logrus does map allocations for every log entry) unless we're using a lot of
its features.

maybe I'm insane, but actually kind of like this... it makes swapping out
logging packages easy if we ever decide to for whatever reason. it's easily
recognized by goimports tool and imported over the stdlib 'log' package and
has all the features that we need and is only 100 LOC and really straight
forward. it's a shame we can't all just agree on a logging api (and Go was so
good at getting rid of many of the bikeshed wars), but this seems like it will
do in a pinch.

one side note, I changed all the references of logrus.WithFields and log15.New
to name their returned variable as 'log', which helps with scoping, as the
package is of the same name, lower scoped 'log' objects will be used
seamlessly when one could easily have accidentally used the package level when
they meant to use the scoped 'log'. it also happens to help when trying to
write sed rules to rewrite all log lines but hopefully we won't have to do
that exercise again.. anyway, it would be nice to maintain this convention in
the future (i.e. be a nazi on PR reviews about it) if we can all agree to it.