dind container uses overlay or aufs now if present
switched to using the docker:1.11-dind image because it was not working to use
the iron/base with the community docker installed later, it was having issues
setting the graph drivers (the point of this). seeing as the base image was
only used to install certs, now this installs the certs, which is a pretty
small surface area to cover and seems reasonable since the docker install
wasn't working.