iron/dind image support btrfs
getting btrfs working is more of a matter of host configuration, from the
runner perspective, this is all that is necessary in order to run on btrfs
storage driver rather than overlay. overlay is causing kernel panics on
coreos. btrfs, so far, is not, and has a relatively negligible performance hit
in testing so far (few seconds added to tasks on saturated instance, no
noticeable difference on unsaturated cluster) -- of course, needs further
testing but we might as well try to commit to it imo.

plenty of info on this and tradeoffs, etc here:
https://docs.docker.com/engine/userguide/storagedriver/btrfs-driver/

it seems like there are more optimization opportunities in the host
configuration that are not perfectly relevant to this patch but that we could
discuss (once we have an ebs drive, how to mount that, what's on it, etc.)