Remove container after updating task status.
This is a short term fix for updating task status first and removing container
next. This is because removing containers can often take a while, sometimes it
can just block due to docker bugs, and we'd like the user to not see this.

I have the verbose defer logic for future-proofness, but happy to just add
removeContainer lines to the 2 spots where we return early and return a nil
runResult if that is more idiomatic.

Long term, we could add a separate removeContainer reaper, with a channel queue
that has a buffer the same size as concurrency or similar, to allow gofers to
continue processing tasks. There are some design concerns around having to make
Docker driver not just have a single entry point of Run(). Stuff like having to
keep that reaper in mind when shutting down the runner, even if all gofers are
ready to shut down and so on. Haven't given it too much thought.