some errors shouldn't error out task
since we are retrying docker endpoints, it means sometimes the previous
request can succeed and the retry failing in certain scenarios is not the end
of the world. found a couple of these, there may be more. I don't think we can
do them in the docker_client wrapper because for things like start container
we'd have to go inspect it and then return, which turns into a big hair ball,
it's much easier to just ignore the returned value in the caller since we
don't need it and do the error filtering there, even though it kind of
clutters up the code a bit :/

an aside, fsouza now appears to have 3 methods for handling errors

............................................________
....................................,.-'"...................``~.,
.............................,.-"..................................."-.,
.........................,/...............................................":,
.....................,?......................................................,
.................../...........................................................,}
................./......................................................,:`^`..}
.............../...................................................,:"........./
..............?.....__.........................................:`.........../
............./__.(....."~-,_..............................,:`........../
.........../(_...."~,_........"~,_....................,:`........_/
..........{.._$;_......"=,_......."-,_.......,.-~-,},.~";/....}
...........((.....*~_......."=-._......";,,./`..../"............../
...,,,___.`~,......"~.,....................`.....}............../
............(....`=-,,.......`........................(......;_,,-"
............/.`~,......`-...................................../
.............`~.*-,.....................................|,./.....,__
,,_..........}.>-._...................................|..............`=~-,
.....`=~-,__......`,.................................
...................`=~-,,.,...............................
................................`:,,...........................`..............__
.....................................`=-,...................,%`>--==``
........................................_..........._,-%.......`
...................................,