error handling yay
the UserVisibleError interface mostly has callers that send it a curated
string and it's easy enough to just log any real errors at that call site,
most invocations passed the same error as the user error and the real error.

remove dockerError... I think this was for the unrecoverable stuff but it
didn't actually do anything but call the embedded errors methods so it is odd
now and thus removed

don't mask any filtered errors from the docker retrying... fuck me, right?

embed sysErrResult errors to maintain other methods

add a new debug log for create container to try to figure out what the error
is that leads to start being called but the container not actually existing
(wtf) -- the expectation is this isn't fixed, but need better debugging

think we may have been masking exit code error messages in the task?