make command parity levels of broken
... somebody change the record

we need to make command a list of arguments at the api level in order to give
the same output as building a Dockerfile and specifying a CMD, our API only
takes a string which means that certain things are interpreted incorrectly.
for example:

["sh", "-c", "echo hello && sleep 60"] -> works as you'd expect
["sh", "-c", "echo hello && sleep 60"] -> doesn't work
["echo hello && sleep 60"] -> doesn't work

in retrospect I think this is a bad example but I'm not sure, however I am
sure:

this means that with certain commands we do not currently give the same output
as you would building a Dockerfile with the same CMD, this is bad but requires
API changes and nobody seems to have complained as of yet (still going to
fix). the current is much more forgiving but imo we should make it same as
docker output