Check for Docker Daemon presence early in initialization
Currently, if the environment is misconfigured and Docker is absent,
then the error about missing Docker presents itself too late, that is,
when there is a task to be ran.

This improves the interface by pinging the daemon during initialization
phase, thus giving the user a hint that something is wrong right away.