move registry crap back into driver
this is now a static method on the docker driver package, as future drivers
may also be interested in talking to the hub and it would be nice to not
conflate the driver and the tasker, since they shouldn't really have to know
about each other. As the CS curricula has endorsed, this does not actually
require 17 parameter functions to get an object that can tell you what you
needed to know. rkt specifically can also use docker images, but we shouldn't
have to configure the docker driver just to check images, this is half of the
reasoning, the other half about toting around a list of driver objects and
checking each of their image checker funcs in the future in the tasker. ew.
seems like registry stuff being in docker driver is useful since a lot of the
helper functions have a lot of overlap, and then people don't have to import
the dockerclient.

this also gets rid of this tasker.Auther thing and then functions can also get
rid of that import. seems like this was meant to supplant passing around
docker creds, which seems like something we might want to do, but for now we
don't really need to worry about that since it's just docker creds and the
taskers can expose different auth mechanisms to each driver.