iron/dind get rid of btrfs
we gave btrfs a pretty good go at it and it's really really slow even though
it doesn't make the kernel panic or run out of inodes (i.e. it works).
overlay2 seems to work with kernel 4.7+ without causing any panics and the
inode issue seems fixed (inodes on our ebs drive with cached images is around
20% usage, about same as we have with aufs -- with 'overlay' we hit 100% inode
usage just doing `docker pull -a iron/images` so it's not really suitable for
our use case). overlay2 has been fast, docker operations for create / start
container back in ms where with btrfs it was taking on the order of minutes
(yes, seriously).

tested this on a coreos box with docker 1.10 and dind 1.12 and ubuntu with
1.12 and dind 1.12 and ubuntu boots up with aufs storage driver and coreos
with overlay2, yay.

I'll push a new `iron/dind` after this is merged.