Make 400 errors from CreateContainer user visible.
The common cause of the container configuration being bad would be due to a bad
image, lack of command/entrypoint, bad env vars and so on. All of which are
controlled by the user.

When we change the container configuration, I'd expect to catch failures due to
that before rolling it out, and even if not, this would still let us know what
went wrong.

Example: https://hud-e.iron.io/worker/projects/5614090bf76a280006000038/tasks/582a7b727582400006fd2fe9