Actually bail out of WaitContainer when context is done. (#16)
The EOF error we were running into would transition to a non-temporary
ErrConnectionRefused (since Daemon is dead). At this point we would get stuck
in the outer loop.

Now, if Docker daemon dies, the task.Timeout context will eventually break the
loop and task can be marked as failed. Of course, this doesn't solve all
problems. I am not sure yet if the AttachToContainerNonBlocking Wait() will
automatically terminate.

Ideally InspectContainer will also start failing in this case, so we can
transition the task to "error".