use context for docker retries (#19)
we're doing some weird straddling of the fsouza api now with contexts, i think
in part because some of the contexts are on options and some are method
parameters. i'm open to ideas, played around with a few things and didn't
really care for any of them. the weirdness is because we want an enclosing
context to have a certain timeout (10m) only if there's no timeout to the
actual fsouza method, and if there's a timeout to the fsouza one then we don't
want that, so it was weird to set up `retry(func())` with the timeout in there
since you'd need it as a parameter to both retry and the function parameter (I
think).

anyway, the benefit is this will get rid of "retrying wait container"
weirdness and that weird logic there. yay