tail writer was panic'ing (#23)
now that wait obeys the task's context the container may still be running when
we go to tail the log (this was previously true if e.g. wait timed out but
container was still running, it just didn't happen enough to notice). tailing
the log is really just a best effort to see what's going on in there, if
another write goes in there we don't really care, but it does bring up the
point that we really should wait until the container exits for real (as we
were, mostly by coincidence and not good coding); i'm not sure how to go about
it since it could end up taking forever and at some point we need to give up,
we could pass a different context to WaitContainerWithContext that has an
extra minute onto it and it would probably cover most cases, though.

we may want to return `0, nil` so that any other writers in the chain won't
fail, either one makes sense but returning nil and not writing something just
feels wrong.